feat: add single prompt sse tests
