+++ test/Cnblogs.DashScope.Sdk.UnitTests/RawHttpData/application-single-generation-text-sse.response.header.txt
@@ -1,0 +1,14 @@
+﻿HTTP/1.1 200 OK
+vary: Origin,Access-Control-Request-Method,Access-Control-Request-Headers
+x-request-id: fec856c6-c146-99bc-940c-fd6c36939a16
+content-type: text/event-stream;charset=UTF-8
+x-dashscope-call-gateway: true
+x-dashscope-timeout: 180
+x-dashscope-finished: false
+req-cost-time: 1755
+req-arrive-time: 1742113457368
+resp-start-time: 1742113459124
+x-envoy-upstream-service-time: 1744
+date: Sun, 16 Mar 2025 08:24:18 GMT
+server: istio-envoy
+transfer-encoding: chunked
