+++ test/Cnblogs.DashScope.Sdk.UnitTests/RawHttpData/application-single-generation-text-with-thought-nosse.request.body.json
@@ -1,0 +1,16 @@
+﻿{
+  "input": {
+    "prompt": "总结xUnit Test Patterns中的内容"
+  },
+  "parameters": {
+    "has_thoughts": true,
+    "top_k": 100,
+    "top_p": 0.8,
+    "seed": 1234,
+    "temperature": 0.85,
+    "rag_options": {
+      "pipeline_ids": ["thie5bysoj"],
+      "file_ids": ["file_d129d632800c45aa9e7421b30561f447_10207234"]
+    }
+  }
+}
