+++ test/Cnblogs.DashScope.Sdk.UnitTests/RawHttpData/application-single-generation-text-with-thought-nosse.request.header.txt
@@ -1,0 +1,8 @@
+﻿POST /api/v1/apps/010aa085cc9943268731861c9511bb0c/completion HTTP/1.1
+Content-Type: application/json
+Accept: */*
+Cache-Control: no-cache
+Host: dashscope.aliyuncs.com
+Accept-Encoding: gzip, deflate, br
+Connection: keep-alive
+Content-Length: 403
