+++ test/Cnblogs.DashScope.Sdk.UnitTests/RawHttpData/application-single-generation-text-with-thought-nosse.response.header.txt
@@ -1,0 +1,15 @@
+﻿HTTP/1.1 200 OK
+vary: Origin,Access-Control-Request-Method,Access-Control-Request-Headers, Accept-Encoding
+content-type: application/json
+x-request-id: b5819020-e5aa-9481-8c9d-e11797f191d8
+x-dashscope-timeout: 180
+x-dashscope-call-gateway: true
+x-dashscope-finished: true
+req-cost-time: 43838
+req-arrive-time: 1742111961733
+resp-start-time: 1742112005571
+x-envoy-upstream-service-time: 43826
+content-encoding: gzip
+date: Sun, 16 Mar 2025 08:00:05 GMT
+server: istio-envoy
+transfer-encoding: chunked
