+++ test/Cnblogs.DashScope.Sdk.UnitTests/Utils/Snapshots.TextGeneration.cs
@@ -1,0 +1,614 @@
+﻿using Cnblogs.DashScope.Core;
+using Json.Schema;
+using Json.Schema.Generation;
+
+namespace Cnblogs.DashScope.Sdk.UnitTests.Utils;
+
+public static partial class Snapshots
+{
+    public static class TextGeneration
+    {
+        public static class TextFormat
+        {
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SinglePrompt = new(
+                    "single-generation-text",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input = new TextGenerationInput { Prompt = "请问 1+1 是多少？" },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "text",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = false
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            FinishReason = "stop", Text = "1+1 等于 2。这是最基本的数学加法之一，在十进制计数体系中，任何情况下两个一相加的结果都是二。"
+                        },
+                        RequestId = "4ef2ed16-4dc3-9083-a723-fb2e80c84d3b",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            InputTokens = 8,
+                            OutputTokens = 35,
+                            TotalTokens = 43
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SinglePromptIncremental = new(
+                    "single-generation-text",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input = new TextGenerationInput { Prompt = "请问 1+1 是多少？" },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "text",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = true
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput { FinishReason = "stop", Text = "1+1等于2。" },
+                        RequestId = "5b441aa7-0b9c-9fbc-ae0a-e2b212b71eac",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            InputTokens = 16,
+                            OutputTokens = 6,
+                            TotalTokens = 22
+                        }
+                    });
+        }
+
+        public static class MessageFormat
+        {
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SingleMessage = new(
+                    "single-generation-message",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input =
+                            new TextGenerationInput { Messages = [TextChatMessage.User("请问 1+1 是多少？")] },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = false
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop",
+                                    Message = TextChatMessage.Assistant(
+                                        "1+1 等于 2。这是最基本的数学加法之一，在十进制计数体系中，任何两个相同的数字相加都等于该数字的二倍。")
+                                }
+                            ]
+                        },
+                        RequestId = "e764bfe3-c0b7-97a0-ae57-cd99e1580960",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 47,
+                            OutputTokens = 39,
+                            InputTokens = 8
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SingleChatClientMessage = new(
+                    "single-generation-message",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input =
+                            new TextGenerationInput { Messages = [TextChatMessage.User("请问 1+1 是多少？")] },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            ToolChoice = ToolChoice.AutoChoice
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop",
+                                    Message = TextChatMessage.Assistant(
+                                        "1+1 等于 2。这是最基本的数学加法之一，在十进制计数体系中，任何两个相同的数字相加都等于该数字的二倍。")
+                                }
+                            ]
+                        },
+                        RequestId = "e764bfe3-c0b7-97a0-ae57-cd99e1580960",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 47,
+                            OutputTokens = 39,
+                            InputTokens = 8
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SingleMessageJson = new(
+                    "single-generation-message-json",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input =
+                            new TextGenerationInput { Messages = [TextChatMessage.User("请问 1+1 是多少？用 JSON 格式输出。")] },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = false,
+                            ResponseFormat = DashScopeResponseFormat.Json
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop",
+                                    Message = TextChatMessage.Assistant("{\n  \"result\": 2\n}")
+                                }
+                            ]
+                        },
+                        RequestId = "6af9571b-1033-98f9-a287-c06f2e9d6f7f",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 34,
+                            OutputTokens = 9,
+                            InputTokens = 25
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SingleMessageIncremental = new(
+                    "single-generation-message",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input =
+                            new TextGenerationInput { Messages = [TextChatMessage.User("请问 1+1 是多少？")] },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = true
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop",
+                                    Message = TextChatMessage.Assistant(
+                                        "1+1 等于 2。这是最基本的数学加法之一，在十进制计数体系中，任何情况下 1 加上另一个 1 的结果都是 2。")
+                                }
+                            ]
+                        },
+                        RequestId = "d272255f-82d7-9cc7-93c5-17ff77024349",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 48,
+                            OutputTokens = 40,
+                            InputTokens = 8
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                SingleMessageChatClientIncremental = new(
+                    "single-generation-message",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input =
+                            new TextGenerationInput { Messages = [TextChatMessage.User("请问 1+1 是多少？")] },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new[] { "你好" },
+                            IncrementalOutput = true,
+                            ToolChoice = ToolChoice.AutoChoice
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop",
+                                    Message = TextChatMessage.Assistant(
+                                        "1+1 等于 2。这是最基本的数学加法之一，在十进制计数体系中，任何情况下 1 加上另一个 1 的结果都是 2。")
+                                }
+                            ]
+                        },
+                        RequestId = "d272255f-82d7-9cc7-93c5-17ff77024349",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 48,
+                            OutputTokens = 40,
+                            InputTokens = 8
+                        }
+                    });
+
+            public static readonly
+                RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>> SingleMessageWithTools =
+                    new(
+                        "single-generation-message-with-tools",
+                        new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                        {
+                            Model = "qwen-max",
+                            Input = new TextGenerationInput { Messages = [TextChatMessage.User("杭州现在的天气如何？")] },
+                            Parameters = new TextGenerationParameters()
+                            {
+                                ResultFormat = "message",
+                                Seed = 1234,
+                                MaxTokens = 1500,
+                                TopP = 0.8f,
+                                TopK = 100,
+                                RepetitionPenalty = 1.1f,
+                                PresencePenalty = 1.2f,
+                                Temperature = 0.85f,
+                                Stop = new TextGenerationStop("你好"),
+                                EnableSearch = false,
+                                IncrementalOutput = false,
+                                Tools =
+                                [
+                                    new ToolDefinition(
+                                        "function",
+                                        new FunctionDefinition(
+                                            "get_current_weather",
+                                            "获取现在的天气",
+                                            new JsonSchemaBuilder().FromType<GetCurrentWeatherParameters>(
+                                                    new SchemaGeneratorConfiguration
+                                                    {
+                                                        PropertyNameResolver = PropertyNameResolvers.LowerSnakeCase
+                                                    })
+                                                .Build()))
+                                ],
+                                ToolChoice = ToolChoice.FunctionChoice("get_current_weather")
+                            }
+                        },
+                        new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                        {
+                            Output = new TextGenerationOutput
+                            {
+                                Choices =
+                                [
+                                    new TextGenerationChoice
+                                    {
+                                        FinishReason = "stop",
+                                        Message = TextChatMessage.Assistant(
+                                            string.Empty,
+                                            toolCalls:
+                                            [
+                                                new ToolCall(
+                                                    "call_cec4c19d27624537b583af",
+                                                    ToolTypes.Function,
+                                                    0,
+                                                    new FunctionCall(
+                                                        "get_current_weather",
+                                                        """{"location": "浙江省杭州市"}"""))
+                                            ])
+                                    }
+                                ]
+                            },
+                            RequestId = "67300049-c108-9987-b1c1-8e0ee2de6b5d",
+                            Usage = new TextGenerationTokenUsage
+                            {
+                                InputTokens = 211,
+                                OutputTokens = 8,
+                                TotalTokens = 219
+                            }
+                        });
+
+            public static readonly
+                RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>> SingleMessageChatClientWithTools =
+                    new(
+                        "single-generation-message-with-tools",
+                        new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                        {
+                            Model = "qwen-max",
+                            Input = new TextGenerationInput { Messages = [TextChatMessage.User("杭州现在的天气如何？")] },
+                            Parameters = new TextGenerationParameters()
+                            {
+                                ResultFormat = "message",
+                                Seed = 1234,
+                                MaxTokens = 1500,
+                                TopP = 0.8f,
+                                TopK = 100,
+                                RepetitionPenalty = 1.1f,
+                                PresencePenalty = 1.2f,
+                                Temperature = 0.85f,
+                                Tools =
+                                [
+                                    new ToolDefinition(
+                                        "function",
+                                        new FunctionDefinition(
+                                            "get_current_weather",
+                                            "获取现在的天气",
+                                            new JsonSchemaBuilder().FromType<GetCurrentWeatherParameters>(
+                                                    new SchemaGeneratorConfiguration
+                                                    {
+                                                        PropertyNameResolver = PropertyNameResolvers.LowerSnakeCase
+                                                    })
+                                                .Build()))
+                                ],
+                                ToolChoice = ToolChoice.FunctionChoice("get_current_weather")
+                            }
+                        },
+                        new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                        {
+                            Output = new TextGenerationOutput
+                            {
+                                Choices =
+                                [
+                                    new TextGenerationChoice
+                                    {
+                                        FinishReason = "stop",
+                                        Message = TextChatMessage.Assistant(
+                                            string.Empty,
+                                            toolCalls:
+                                            [
+                                                new ToolCall(
+                                                    "call_cec4c19d27624537b583af",
+                                                    ToolTypes.Function,
+                                                    0,
+                                                    new FunctionCall(
+                                                        "get_current_weather",
+                                                        """{"location": "浙江省杭州市"}"""))
+                                            ])
+                                    }
+                                ]
+                            },
+                            RequestId = "67300049-c108-9987-b1c1-8e0ee2de6b5d",
+                            Usage = new TextGenerationTokenUsage
+                            {
+                                InputTokens = 211,
+                                OutputTokens = 8,
+                                TotalTokens = 219
+                            }
+                        });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                ConversationPartialMessageNoSse = new(
+                    "conversation-generation-message-partial",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>()
+                    {
+                        Model = "qwen-max",
+                        Input = new TextGenerationInput()
+                        {
+                            Messages =
+                            [
+                                TextChatMessage.User("请对“春天来了，大地”这句话进行续写，来表达春天的美好和作者的喜悦之情"),
+                                TextChatMessage.Assistant("春天来了，大地", true)
+                            ]
+                        },
+                        Parameters = new TextGenerationParameters()
+                        {
+                            ResultFormat = ResultFormats.Message,
+                            Seed = 1234,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>()
+                    {
+                        RequestId = "4c45d7fd-3158-9ff4-96a0-6e92c710df2c",
+                        Output = new TextGenerationOutput()
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice()
+                                {
+                                    FinishReason = "stop",
+                                    Message =
+                                        TextChatMessage.Assistant(
+                                            "仿佛从漫长的冬眠中苏醒过来，万物复苏。嫩绿的小草悄悄地探出了头，争先恐后地想要沐浴在温暖的阳光下；五彩斑斓的花朵也不甘示弱，竞相绽放着自己最美丽的姿态，将田野、山林装扮得分外妖娆。微风轻轻吹过，带来了泥土的气息与花香混合的独特香味，让人心旷神怡。小鸟们开始忙碌起来，在枝头欢快地歌唱，似乎也在庆祝这个充满希望的新季节的到来。这一切美好景象不仅让人感受到了大自然的魅力所在，更激发了人们对生活无限热爱和向往的心情。")
+                                }
+                            ]
+                        },
+                        Usage = new TextGenerationTokenUsage()
+                        {
+                            TotalTokens = 165,
+                            OutputTokens = 131,
+                            InputTokens = 34
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                ConversationMessageIncremental = new(
+                    "conversation-generation-message",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-max",
+                        Input =
+                            new TextGenerationInput
+                            {
+                                Messages =
+                                [
+                                    TextChatMessage.User("现在请你记住一个数字，42"),
+                                    TextChatMessage.Assistant("好的，我已经记住了这个数字。"),
+                                    TextChatMessage.User("请问我刚才提到的数字是多少？")
+                                ]
+                            },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = true
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop", Message = TextChatMessage.Assistant("您刚才提到的数字是42。")
+                                }
+                            ]
+                        },
+                        RequestId = "9188e907-56c2-9849-97f6-23f130f7fed7",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 33,
+                            OutputTokens = 9,
+                            InputTokens = 24
+                        }
+                    });
+
+            public static readonly RequestSnapshot<ModelRequest<TextGenerationInput, ITextGenerationParameters>,
+                    ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>>
+                ConversationMessageWithFilesIncremental = new(
+                    "conversation-generation-message-with-files",
+                    new ModelRequest<TextGenerationInput, ITextGenerationParameters>
+                    {
+                        Model = "qwen-long",
+                        Input =
+                            new TextGenerationInput
+                            {
+                                Messages =
+                                [
+                                    TextChatMessage.File(
+                                        ["file-fe-WTTG89tIUTd4ByqP3K48R3bn", "file-fe-l92iyRvJm9vHCCfonLckf1o2"]),
+                                    TextChatMessage.User("这两个文件是相同的吗？")
+                                ]
+                            },
+                        Parameters = new TextGenerationParameters
+                        {
+                            ResultFormat = "message",
+                            Seed = 1234,
+                            MaxTokens = 1500,
+                            TopP = 0.8f,
+                            TopK = 100,
+                            RepetitionPenalty = 1.1f,
+                            Temperature = 0.85f,
+                            Stop = new int[][] { [37763, 367] },
+                            EnableSearch = false,
+                            IncrementalOutput = true
+                        }
+                    },
+                    new ModelResponse<TextGenerationOutput, TextGenerationTokenUsage>
+                    {
+                        Output = new TextGenerationOutput
+                        {
+                            Choices =
+                            [
+                                new TextGenerationChoice
+                                {
+                                    FinishReason = "stop",
+                                    Message = TextChatMessage.Assistant(
+                                        "你上传的两个文件并不相同。第一个文件`test1.txt`包含两行文本，每行都是“测试”。而第二个文件`test2.txt`只有一行文本，“测试2”。尽管它们都含有“测试”这个词，但具体内容和结构不同。")
+                                }
+                            ]
+                        },
+                        RequestId = "7865ae43-8379-9c79-bef6-95050868bc52",
+                        Usage = new TextGenerationTokenUsage
+                        {
+                            TotalTokens = 115,
+                            OutputTokens = 57,
+                            InputTokens = 58
+                        }
+                    });
+        }
+    }
+}
